fix(backend): better calculation of summary

+++ apps/backend/src/models.rs
@@ -781,7 +781,7 @@ pub mod media {
     )]
     pub struct UserFitnessWorkoutSummary {
         pub recorded: u64,
-        pub duration: u64,
+        pub duration: Decimal,
         pub weight: Decimal,
     }
